nocloud: create temp mount dir under /run
The /run is usually a tmpfs which is automatically removed on reboot.
Use this for the temp mount dir so we don't leave any /mnt/cidata-* temp
leftovers.

Also explicitly create the parent directory of the temp dir, which may
no exist.  This solved a weird case where running tests locally ended up
hanging in mount.